pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "neuroai-project"
version = "0.1.0"
description = "Starter scaffold for NeuroAI and industrial visual anomaly detection experiments."
readme = "README.md"
requires-python = ">=3.10,<3.13"
dependencies = [
    "numpy",
    "pandas",
    "matplotlib",
    "scikit-learn",
    "pyyaml",
    "tqdm",
    "torch",
    "torchvision",
    "torchaudio",
    "gradio",
    "spikingjelly",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "ruff",
    "ipykernel",
    "jupyter",
]
event = ["tonic"]
export = [
    "onnx",
    "onnxruntime",
]

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.ruff]
target-version = "py310"
line-length = 100
src = ["src", "demo", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]

[tool.pytest.ini_options]
addopts = "-q"
pythonpath = ["."]
testpaths = ["tests"]
